+++ src/app/shared/shared.module.ts
@@ -1,21 +1,36 @@
+import { TranslateHttpLoader } from '@ngx-translate/http-loader';
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
-import { TranslateModule } from '@ngx-translate/core';
+import { TranslateModule, TranslateLoader } from '@ngx-translate/core';
 
 import { SearchBarComponent } from './search-bar/search-bar.component';
+import { HttpClient, HttpClientModule } from '@angular/common/http';
+
+export function createTranslateLoader(http: HttpClient) {
+  return new TranslateHttpLoader(http, './assets/i18n/', '.json');
+}
 
 @NgModule({
   imports: [
     CommonModule,
     FormsModule,
-    TranslateModule,
+    HttpClientModule,
+    TranslateModule.forRoot({
+      loader: {
+        provide: TranslateLoader,
+        useFactory: (createTranslateLoader),
+        deps: [HttpClient]
+      },
+      useDefaultLang: true,
+    }),
   ],
   declarations: [
     SearchBarComponent,
   ],
   exports: [
     SearchBarComponent,
+    TranslateModule
   ]
 })
 export class SharedModule { }
